pyproject.toml
[tool.poetry]
name = "rankkeeper"
version = "0.1.0"
description = ""
authors = ["kousakirai <kouki.muro2850@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.8.10"
discord-py = "^2.3.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
